+++ quiltcore/config.py
@@ -3,15 +3,15 @@ from pathlib import Path
 from un_yaml import UnYaml  # type: ignore
 
 
-class CoreConfig(UnYaml):
-    CONFIG_FILE = "quiltcore.yaml"
+class Config(UnYaml):
+    CONFIG_FILE = "yaml/quiltcore.yaml"
 
     @classmethod
     def DefaultConfig(cls) -> dict:
         return UnYaml.LoadYaml(cls.CONFIG_FILE, __package__)
 
     def __init__(self, yaml_data: dict = {}) -> None:
-        data = yaml_data if len(yaml_data) > 0 else CoreConfig.DefaultConfig()
+        data = yaml_data if len(yaml_data) > 0 else Config.DefaultConfig()
         super().__init__(data)
 
     def get_str(self, key: str, default="") -> str: