+++ quiltcore/manifest.py
@@ -6,30 +6,36 @@ import pyarrow.json as pj  # type: ignore
 from typing_extensions import Self
 from upath import UPath
 
-from .resource import CoreResource
+from .resource import Resource
 
 
-class CoreManifest(CoreResource):
-    """In-memory representation of a serialized package manifest."""
+class Manifest(Resource):
+    """
+    In-memory representation of a serialized package manifest.
+    list/get returns Blob with Path to the Place data actually lives
+    """
 
     # TODO: cache Blobs to avoid repeated lookups
-    # TODO: improve/replace child(key) handling
 
     def __init__(self, path: Path, **kwargs):
         super().__init__(path, **kwargs)
         with path.open(mode="rb") as fi:
             self.table = pj.read_json(fi)
-        self.body = self.setup()
+        self.body = self.setup_table()
         self.name_col = self.cf.get_str("schema/name", "logical_key")
-        self.loc_col = self.cf.get_str("schema/location", "physical_keys")
+        self.place_col = self.cf.get_str("schema/place", "physical_keys")
 
-    def setup(self) -> pa.Table:
+    def setup_table(self) -> pa.Table:
         first = self.table.take([0]).to_pydict()
         headers = self.cf.get_dict("schema/headers")
         keys = list(headers.keys())
         for key in keys:
             setattr(self, key, first[key][0])
         return self.table.drop_columns(keys).slice(1)
+
+    #
+    # Private Methods for child resources
+    # 
 
     def child_row(self, key: str) -> dict:
         """Return the dict for a child resource."""
@@ -40,15 +46,19 @@ class CoreManifest(CoreResource):
         return rows.to_pydict()
 
     def child_path(self, key: str) -> Path:
-        """Return the path for a child resource."""
+        """Return the Path for a child resource."""
         row = self.child_row(key)
-        loc = row[self.loc_col][0][0]
-        return UPath(loc)
+        place = row[self.place_col][0][0]
+        return UPath(place)
 
     def child_args(self, key: str) -> dict:
         """Return the parameters for a child resource."""
         row = self.child_row(key)
         return {"row": row, "parent": self}
+
+    #
+    # Public HTTP-like Methods
+    # 
 
     def list(self) -> list[Self]:
         """List all child resources."""