+++ quiltcore/namespace.py
@@ -1,14 +1,17 @@
 from pathlib import Path
 
-from .resource import CoreResource
+from .resource import Resource
 
 
-class CoreName(CoreResource):
-    """Namespace of Manifests by Hash"""
+class Namespace(Resource):
+    """
+    Namespacespace of Manifests by Hash
+    list/get returns a specific Manifest
+    """
 
     def __init__(self, path: Path, **kwargs):
         super().__init__(path, **kwargs)
-        self.values = kwargs["values"]
+        self.versions = kwargs["versions"]
 
     def hash(self, tag: str = "latest") -> str:
         hash_file = self.path / tag
@@ -17,4 +20,4 @@ class CoreName(CoreResource):
     def child_path(self, key: str) -> Path:
         """Return the path for a child resource."""
         hash = self.hash(key)
-        return self.values.path / hash
+        return self.versions / hash