+++ quiltcore/registry.py
@@ -1,21 +1,20 @@
 from pathlib import Path
 
-from .resource import CoreResource
-from .values import CoreValues
+from .resource import Resource
 
 
-class CoreRegistry(CoreResource):
+class Registry(Resource):
     """
-    Registry of Names and parent of Values
-    `list` and `get` return CoreName with parent of CoreValues
+    Top-level Resource reperesenting a Quilt Registry.
+    Defines core paths containing Namespaces and Manifests.
+    `list` and `get` return Namespace objects
     """
 
     def __init__(self, path: Path, **kwargs):
         super().__init__(path, **kwargs)
         base = path / self.cf.get_path("dirs/config")
         self.path = base / self.cf.get_path("dirs/names")
-        values = base / self.cf.get_path("dirs/values")
-        self.values = CoreValues(values, self)
+        self.versions = base / self.cf.get_path("dirs/versions")
 
     def child_args(self, key: str) -> dict:
-        return {"values": self.values}
+        return {"versions": self.versions}