+++ tests/test_blob.py
@@ -1,7 +1,7 @@
 from tempfile import TemporaryDirectory
 
 from pytest import fixture
-from quiltcore import CoreBlob, CoreManifest
+from quiltcore import Blob, Manifest
 from upath import UPath
 
 from .conftest import TEST_KEY, TEST_OBJ, TEST_OBJ_HASH, TEST_TABLE
@@ -14,13 +14,13 @@ HASH_HW = "c0535e4be2b79ffd93291305436bf889314e4a3faec05ecffcbb7df31ad9e51a"
 @fixture
 def blob():
     path = UPath(TEST_TABLE)
-    man = CoreManifest(path)
-    blob: CoreBlob = man.get(TEST_KEY)  # type: ignore
+    man = Manifest(path)
+    blob: Blob = man.get(TEST_KEY)  # type: ignore
     return blob
 
 
-def test_blob(blob: CoreBlob):
-    assert isinstance(blob.parent, CoreManifest)
+def test_blob(blob: Blob):
+    assert isinstance(blob.parent, Manifest)
     assert blob.logical_key == TEST_KEY  # type: ignore
     assert blob.physical_key == TEST_OBJ  # type: ignore
     assert blob.hash["value"] == TEST_OBJ_HASH  # type: ignore
@@ -36,7 +36,7 @@ def test_blob(blob: CoreBlob):
     assert meta["target"] == "parquet"
 
 
-def test_blob_put(blob: CoreBlob):
+def test_blob_put(blob: Blob):
     with TemporaryDirectory() as tmpdirname:
         dest = UPath(tmpdirname) / TEST_KEY
         loc = blob.put(dest)
@@ -45,21 +45,21 @@ def test_blob_put(blob: CoreBlob):
         assert loc.exists()
 
 
-def test_blob_digest(blob: CoreBlob):
+def test_blob_digest(blob: Blob):
     assert blob.hash_digest
     digest = blob.hash_digest.digest(DATA_HW)
-    assert digest.hex() == CoreBlob.MH_PREFIX + HASH_HW
+    assert digest.hex() == Blob.MH_PREFIX_SHA256 + HASH_HW
 
     digest2 = blob.digest(DATA_HW)
     assert digest2 == HASH_HW
 
 
-def test_blob_digest_verify(blob: CoreBlob):
+def test_blob_digest_verify(blob: Blob):
     blob.hash_value = HASH_HW
     assert blob.verify(DATA_HW)
 
 
-def test_blob_verify(blob: CoreBlob):
+def test_blob_verify(blob: Blob):
     assert blob.hash_value
     with TemporaryDirectory() as tmpdirname:
         dest = UPath(tmpdirname) / TEST_KEY