+++ tests/test_registry.py
@@ -1,5 +1,5 @@
 from pytest import fixture, raises
-from quiltcore import CoreManifest, CoreName, CoreRegistry
+from quiltcore import Manifest, Namespace, Registry
 from upath import UPath
 
 from .conftest import TEST_BKT, TEST_HASH, TEST_PKG, TEST_TAG
@@ -8,15 +8,15 @@ from .conftest import TEST_BKT, TEST_HASH, TEST_PKG, TEST_TAG
 @fixture
 def reg():
     path_bkt = UPath(TEST_BKT)
-    return CoreRegistry(path_bkt)
+    return Registry(path_bkt)
 
 
 def test_registry(reg):
     assert reg
     assert reg.cf
     assert ".quilt/named_packages" in str(reg.path)
-    assert ".quilt/packages" in str(reg.values)
-    assert reg.values.path.exists()
+    assert ".quilt/packages" in str(reg.versions)
+    assert reg.versions.exists()
     assert reg.path.is_dir()
 
 
@@ -25,13 +25,13 @@ def test_registry_list(reg):
     assert isinstance(result, list)
     assert len(result) > 0
     first = result[0]
-    assert isinstance(first, CoreName)
+    assert isinstance(first, Namespace)
     assert TEST_PKG in str(first)
 
 
 def test_registry_get(reg):
     name = reg.get(TEST_PKG)
-    assert isinstance(name, CoreName)
+    assert isinstance(name, Namespace)
     assert TEST_PKG in str(name)
 
     with raises(KeyError):
@@ -41,8 +41,8 @@ def test_registry_get(reg):
 def test_name_get(reg):
     name = reg.get(TEST_PKG)
     man = name.get(TEST_TAG)
-    assert isinstance(man, CoreManifest)
+    assert isinstance(man, Manifest)
     assert TEST_HASH in str(man)
 
     latest = name.get("latest")
-    assert isinstance(latest, CoreManifest)
+    assert isinstance(latest, Manifest)